Capitalize file names
When the capitalization of these files changed, `git` did not pick up
the change as in Mac core.ignorecase is true by default in MacOS git
installations.  As a result, the files were checked in with wrong case,
which does not work on sane (case-sensitive) filesystems.

Fix that by setting `git config core.ignorecase false` and checking in
again.